DateTimePicker typesafe, missing fixing icon
