+++ components/Link-Generator/Date-time/CalendarIcon.tsx
@@ -1,0 +1,20 @@
+// CalendarIcon.tsx
+
+export const CalendarIcon = (props: React.SVGProps<SVGSVGElement>) => (
+	<svg
+		{...props}
+		xmlns="http://www.w3.org/2000/svg"
+		fill="none"
+		viewBox="0 0 24 24"
+		stroke="currentColor"
+		strokeWidth={2}
+		aria-labelledby="calendarIconTitle"
+	>
+		<title id="calendarIconTitle">Open Calendar</title>
+		<path
+			strokeLinecap="round"
+			strokeLinejoin="round"
+			d="M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z"
+		/>
+	</svg>
+);
